+++ scan/scan_longn.c
@@ -6,7 +6,7 @@ size_t scan_longn(const char *src,size_t n,long *dest) {
   register const char *tmp;
   register long int l;
   register unsigned char c;
-  int neg;
+  unsigned int neg;
   int ok;
   if (!n--) return 0;
   tmp=src; l=0; ok=neg=0;
@@ -24,12 +24,12 @@ size_t scan_longn(const char *src,size_t n,long *dest) {
      * so instead of *10 we do (l<<3) (i.e. *8) + (l<<1) (i.e. *2)
      * and check for overflow on all the intermediate steps */
     n=(unsigned long)l<<3; if ((n>>3)!=(unsigned long)l) break;
-    if (n+(l<<1) < n) break;
-    n+=l<<1;
+    if (n+((unsigned long)l<<1) < n) break;
+    n+=(unsigned long)l<<1;
     if (n+c < n) break;
     n+=c;
     if (n > maxlong+neg) break;
-    l=n;
+    l=(long)n;
     ++tmp;
     ok=1;
   }
