+++ scan/scan_nonwhitenskip.c
@@ -5,5 +5,5 @@ size_t scan_nonwhitenskip(const char *s,size_t limit) {
   register const char *t=s;
   register const char *u=t+limit;
   while (t<u && !isspace(*t)) ++t;
-  return t-s;
+  return (size_t)(t-s);
 }
